services/stt/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai-whisper==20231117
pydub==0.25.1
numpy==1.26.4
